refactor: clean file structure
