+++ src/pages/BoardPage/ui/Field/Skeleton/ColumnSkeleton.tsx
