+++ src/pages/BoardPage/ui/Field/Skeleton/FieldSkeleton.tsx
