+++ src/pages/BoardPage/ui/Modals/DelColModal.tsx
