+++ src/pages/BoardPage/ui/Panel/Panel.tsx
